autocomp: append spaces to successful completion hints
The completion is assumed successful:
- when there remains only one completion
  state
- this completion state does not represent
  a node that has children

On successful completion, appending a space to this
unique GNU readline completion state will tell the user
pressing TAB again is useless.

Note that when a node's children enumeration could
subsequently happen, TAB completion is prepared by instead
appending a trailing '/' to the unique hint.