dtsh: explicit missing bindings issue
The current implementation of the devicetree model initialization
may filter out bindings that never appear first (i.e. as most specific)
in the "compatible" list of a node.
For example, the binding for "nordic,nrf-swi" is likely to
always be masked by a more specific compatible, e.g. "nordic,nrf-egu".

See:
- Dtsh.dt_bindings()
- Dtsh._init_bindings()