tui: in node table, match register(s) column to unit-address
The 'Registers' column for the node tables enumerates
the 'reg' property for the node: this is not the most
appropriate value for the 'r' node table format specifier.

We think the 'r' column value should instead show
the register that matches the node's unit-address.

Note: accessing all registers for a node remains possible
with the 'cat' built-in.