ci: patch upstream python-devicetree library
The PropertySec.path API in the python-devicetree library
wrongly tells all properties originate from the top-level
binding file, which causes the cat command to show
misleading and confusing paths (see issue #5).

We've proposed a patch upstream, but its review clearly is
not a priority (the underlying issue has no consequence on
Zephyr use of the library).

Applying this patch here:
- does actually fix the cat command, and does not otherwise break DTSh
- can't break Zephyr in any way (will still use its own unpatched
  version of the python-devicetree library, see issue #2)