rich: io: fix record/capture approach for output redirection
When redirecting commands output to files, we need to:
- record the output as rich segments we'll later
  format to text, HTML or SVG
- capture the ANSI output that would otherwise be echoed
  to the console

Our approach for mixing record and capture has proven
to be inappropriate, and incompatible with latest
versions of the rich library.

The recommened approach for a throwaway capture stream
is now to open the console with a StringIO instance as stdout.

Unpining the rich library version since we don't need
this workaround anymore.

Fixes: #7