rich: io: disabling markup in all Console objects
Rich markup (similar to bbcode) is enabled by default
in Console instances, and square brackets are thus interpreted.

For example, writing the string bellow to the console:

    "include: [cortex-m-systick.yaml]"

would just render as:

      "include:"

This is not an issue when using what we call the "long list format",
i.e. when writing rich outputs (typically with "-l"): in this case,
we don't directly write strings that may be interpreted
but instances of rich Renderable.

When using raw ASCII (typically anything without the "-l" flag),
we do write strings to the Console, and the interpreter may eat
some pieces of the output.
This shows e.g. when dumping the content of a YAML file with "cat -Y".

Disabling markup support in consoles (we don't rely on it).