rich: svg: fix confusing pylint E1101
Pylint seems confused by factory methods that return tuples:

    class Fragment:
        @classmethod
        def mk_fragment(cls) -> Tuple[int, "Fragment"]:
            return 0, cls()

        def foobar(self) -> None:
            pass

    class Document:
        fragment: Fragment

        def __init__(self) -> None:
            f: Fragment

            _, f = Fragment.mk_fragment()
            f.foobar()

            self.fragment = f
            # E1101: Instance of 'Document' has no 'foobar' member.
            self.fragment.foobar()

            _, self.fragment = Fragment.mk_fragment()
            # E1101: Instance of 'Document' has no 'foobar' member.
            self.fragment.foobar()

What's really confusing is that Pylint complains only when
we make a class member from the Fragment objet,
not when directly using the return value as in f.foobar().
Also note that:
- Pylint is fine if mk_fragment() returns cls() instead of a tuple
- MyPy does not complain

We'll try to sort this out with the help of the Pylint community.

Meanwhile, refactor the SVGFragment API to remove these confusing
error messages.