rich: svg: workaround truncated last captured output line
Workaround upstream rich library issue 3576:
Underscores on the last captured line are clipped
when exporting to SVG.

The main <clipPath> height computed
by the rich library in Console.export_svg() is a bit too small,
and the bottom of the last captured output line may be truncated,
depending on the font and the printed characters (e.g. underscores,
but also the bottom of the letter 'y').

This workaround adds the bottom padding to the height computed
by the rich library.

Ref:
- https://github.com/dottspina/dtsh/issues/8
- https://github.com/Textualize/rich/issues/3576