rich: html: helpers API for redirecting commands output to HTML
Provides a simple model layer between the rich library
Console.export_html() API and DTSh commands output redirection
to HTML files.

Although the primary rationale for this module is to setup
an implementation of redirection to HTML files that does not
rely on the inline CSS styles generated by the rich library,
it should also help in the long term to maintain
and evolve our HTML format.